pyproject.toml
[project]
name = "taxi_prediction"
description = "『先輩データサイエンティストからの指南書』（技術評論社）のサンプルリポジトリ"
authors = [
    { name = "Junki Asano" },
    { name = "Toma Tanaka" },
    { name = "Katsuhiro Muto" },
    { name = "Masaya Kimura" },
    { name = "Mizuho Yanagi" },
]
version = "1.0.0"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12,<3.13"
dependencies = [
    "chardet>=5.2.0",
    "google-cloud-bigquery>=3.32.0",
    "lightgbm>=4.6.0",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "scikit-learn>=1.5.2",
    "statsmodels>=0.14.4",
    "streamlit>=1.45.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "hydra-core>=1.3.2",
    "japanize-matplotlib>=1.1.3",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "mlflow>=2.22.0",
    "mypy>=1.15.0",
    "pandera>=0.23.1",
    "pytest>=8.3.5",
    "ruff>=0.11.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
# 有効化するルール
# cf. https://beta.ruff.rs/docs/rules/
select = [
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # Pyflakes
    "I",  # isort
    "W",  # pycodestyle warning
]
# 自動的に修正するルール
fixable = ["E", "F", "I"]
# 自動的に修正しないルール
unfixable = ["W"]

[tool.mypy]
python_version = "3.12"
# チェック項目
warn_unreachable = true  # 到達不可能なコードがある場合に警告
disallow_untyped_defs = true  # 型アノテーションのない関数を許可しない
warn_return_any = true  # 戻り値の型が不明な場合に警告
ignore_missing_imports = true  # サードパーティー製のライブラリを無視
# 表示関係
pretty = true
show_error_context = true

[tool.pytest]
minversion = "6.0"
addopts = "-svv --tb=short --capture=no --full-trace"
testpaths = ["tests/*"]

[tool.pytest.ini_options]
# importlib import mode
# cf. https://docs.pytest.org/en/latest/explanation/goodpractices.html#choosing-a-test-layout
addopts = [
    "--import-mode=importlib",
]
